+++ server/templates/tools/map.html
@@ -27,7 +27,8 @@
   <link rel="stylesheet" href={{url_for('static', filename='css/map.min.css')}} >
   <link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons">
   <link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons+Outlined">
-{% endblock %}
+  <link rel="stylesheet" href="https://npmcdn.com/leaflet@1.0.0-rc.2/dist/leaflet.css" />
+  {% endblock %}
 
 {% block content %}
   <div id="main-pane"></div>
