+++ static/css/tools/map.scss
@@ -294,3 +294,11 @@
 #map-chart {
   padding: 0;
 }
+
+#chart-container .leaflet-container {
+  background: white;
+}
+
+.leaflet-map-container #chart-container {
+  height: 500px;
+}
