+++ static/package.json
@@ -40,9 +40,12 @@
     "d3-regression": "^1.2.8",
     "file-loader": "^6.1.0",
     "geo-albers-usa-territories": "0.1.0",
+    "geoblaze": "^1.1.0",
+    "georaster-layer-for-leaflet": "^3.10.0",
     "html-webpack-plugin": "^4.4.1",
     "js-cookie": "^2.2.1",
     "jszip": "^3.10.0",
+    "leaflet": "^1.9.1",
     "lodash": "^4.17.21",
     "moment": "^2.29.2",
     "pako": "^1.0.11",
@@ -58,12 +61,12 @@
     "resize-observer-polyfill": "^1.5.1",
     "sass": "^1.3.0",
     "sass-loader": "^8.0.2",
+    "speed-measure-webpack-plugin": "^1.5.0",
     "ts-loader": "^9.3.1",
     "typescript": "^4.7.4",
     "webpack": "^5.73.0",
     "webpack-cli": "^4.10.0",
-    "webpack-remove-empty-scripts": "^0.6.3",
-    "speed-measure-webpack-plugin": "^1.5.0"
+    "webpack-remove-empty-scripts": "^0.6.3"
   },
   "devDependencies": {
     "@babel/polyfill": "^7.11.5",
@@ -72,6 +75,7 @@
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^10.4.9",
     "@types/jest": "^28.1.3",
+    "@types/node": "^18.8.3",
     "@types/react": "^16.9.49",
     "@types/react-csv": "^1.1.3",
     "@types/react-table": "^7.7.12",
