+++ static/tsconfig.json
@@ -9,12 +9,13 @@
     "outDir": "./dist/",
     "sourceMap": true,
     "noImplicitAny": false,
-    "module": "commonjs",
+    "module": "ESNext",
     "target": "es5",
     "jsx": "react",
     "allowJs": true,
     "esModuleInterop": true,
     "resolveJsonModule": true,
+    "moduleResolution": "node"
   },
   "exclude": ["node_modules"]
 }
